+++ malcolm/modules/pmac/parts/brickpart.py
@@ -5,9 +5,9 @@ from ..infos import ControllerInfo
 
 
 class BrickPart(builtin.parts.ChildPart):
-    def __init__(self, name, mri):
-        # type: (builtin.parts.APartName, builtin.parts.AMri) -> None
-        super(BrickPart, self).__init__(name, mri, initial_visibility=True)
+    def __init__(self, name, mri, initial_visibility=True):
+        # type: (builtin.parts.APartName, builtin.parts.AMri, builtin.parts.AInitialVisibility) -> None
+        super(BrickPart, self).__init__(name, mri, initial_visibility)
         # Hooks
         self.register_hooked(scanning.hooks.ReportStatusHook,
                              self.report_status)